+++ src/lib_protocol_environment/tezos_protocol_environment.ml
@@ -62,6 +62,7 @@ module V9 = Environment_V9
 module V10 = Environment_V10
 module V11 = Environment_V11
 module V12 = Environment_V12
+module V13 = Environment_V13
 module Memory_context = Memory_context
 module Proxy_context = Proxy_context
 module Proxy_delegate = Proxy_delegate